+++ test_cl/issue-12.cpp
@@ -1,0 +1,10 @@
+#pragma warning(default : 4628)
+#define FOO ::foo
+
+class foo {
+};
+
+template <typename T> class bar {
+};
+
+typedef bar<FOO> buzz;